factor useWindowSize into own file
